Issue #OS-191 Refactoring filterOperator
