+++ java/pojos/src/main/java/io/opensaber/pojos/FilterOperator.java
@@ -1,0 +1,7 @@
+package io.opensaber.pojos;
+
+public enum FilterOperator{
+    eq, or,                                                                    // for object
+    lt, gt, gte, lte, between,                                                 // for object
+    contains, startsWith, endsWith, notContains, notStartsWith, notEndsWith    // for Strings value
+}
