+++ java/registry/src/main/java/io/opensaber/registry/dao/SearchDaoImpl.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 import io.opensaber.pojos.Filter;
-import io.opensaber.pojos.FilterOperators.FilterOperator;
+import io.opensaber.pojos.FilterOperator;
 import io.opensaber.pojos.SearchQuery;
 import io.opensaber.registry.middleware.util.Constants;
 import io.opensaber.registry.util.ReadConfigurator;
@@ -29,9 +29,8 @@ public class SearchDaoImpl implements SearchDao {
 		GraphTraversalSource dbGraphTraversalSource = graphFromStore.traversal().clone();
 		List<Filter> filterList = searchQuery.getFilters();
 		GraphTraversal<Vertex, Vertex> resultGraphTraversal = dbGraphTraversalSource.clone().V().hasLabel(searchQuery.getRootLabel())
-		        .range(searchQuery.getOffset(), 100).limit(searchQuery.getLimit()); //range max value is kept to 100(equal to default limit)
+		        .range(searchQuery.getOffset(), 100).limit(searchQuery.getLimit()); //range max value is equal to default limit(100)
 
-		List<P> predicates = new ArrayList<>();
 		// Ensure the root label is correct
 		if (filterList != null) {
 			for (Filter filter : filterList) {
@@ -40,13 +39,10 @@ public class SearchDaoImpl implements SearchDao {
 				FilterOperator operator = filter.getOperator();
 				String path = filter.getPath();
 
-				//List valueList = getValueList(value);
-
                 switch (operator) {
                 case eq:
                     resultGraphTraversal = resultGraphTraversal.has(property, P.eq(genericValue));
-                    break;
-                    
+                    break;                   
                 case gt:
                     resultGraphTraversal = resultGraphTraversal.has(property, P.gt(genericValue));
                     break;
